+++ Packages/ai.natml.vision.mobilenet-v2/Runtime/MobileNetv2Predictor.cs
@@ -6,6 +6,7 @@
 namespace NatML.Vision {
 
     using System;
+    using System.Threading.Tasks;
     using NatML.Features;
     using NatML.Internal;
     using NatML.Types;
@@ -14,36 +15,44 @@ namespace NatML.Vision {
     /// MobileNet v2 classification predictor.
     /// This predictor classifies an image with the ImageNet labels.
     /// </summary>
-    public sealed class MobileNetv2Predictor : IMLPredictor<(string label, float confidence)> {
+    public sealed class MobileNetv2Predictor : IMLPredictor<MobileNetv2Predictor.Label> {
 
         #region --Client API--
         /// <summary>
-        /// Create the MobileNet v2 classification predictor.
+        /// Class label with confidence score.
         /// </summary>
-        /// <param name="model">MobileNet v2 model.</param>
-        public MobileNetv2Predictor (MLEdgeModel model) => this.model = model as MLEdgeModel;
+        public struct Label {
+
+            /// <summary>
+            /// Class label.
+            /// </summary>
+            public string label;
+
+            /// <summary>
+            /// Unnormalized confidence score.
+            /// </summary>
+            public float confidence;
+
+            public void Deconstruct (out string label, out float confidence) {
+                label = this.label;
+                confidence = this.confidence;
+            }
+        }
 
         /// <summary>
         /// Classify an image.
         /// </summary>
         /// <param name="inputs">Input image feature.</param>
         /// <returns>Output label with unnormalized confidence value.</returns>
-        public (string label, float confidence) Predict (params MLFeature[] inputs) {
-            // Check
-            if (inputs.Length != 1)
-                throw new ArgumentException(@"MobileNet v2 predictor expects a single feature", nameof(inputs));
-            // Check type
-            var input = inputs[0];
-            if (!MLImageType.FromType(input.type))
-                throw new ArgumentException(@"MobileNet v2 predictor expects an an array or image feature", nameof(inputs));
+        public Label Predict (params MLFeature[] inputs) {
             // Apply image pre-processing
+            var input = inputs[0];
             if (input is MLImageFeature imageFeature) {
                 (imageFeature.mean, imageFeature.std) = model.normalization;
                 imageFeature.aspectMode = model.aspectMode;
             }
             // Predict
-            var inputType = model.inputs[0];
-            using var inputFeature = (input as IMLEdgeFeature).Create(inputType);
+            using var inputFeature = (input as IMLEdgeFeature).Create(model.inputs[0]);
             using var outputFeatures = model.Predict(inputFeature);
             // Find label
             var logits = new MLArrayFeature<float>(outputFeatures[0]);
@@ -51,8 +60,27 @@ namespace NatML.Vision {
             for (int i = 1, ilen = logits.shape[1]; i < ilen; ++i)
                 argMax = logits[0,i] > logits[0,argMax] ? i : argMax;
             // Return
-            var result = (model.labels[argMax], logits[argMax]);
+            var result = new Label { label = model.labels[argMax], confidence = logits[argMax] };
             return result;
+        }
+
+        /// <summary>
+        /// Dispose the predictor and release resources.
+        /// </summary>
+        public void Dispose () => model.Dispose();
+
+        /// <summary>
+        /// Create the MobileNet v2 predictor.
+        /// </summary>
+        /// <param name="configuration">Edge model configuration.</param>
+        /// <param name="accessKey">NatML access key.</param>
+        public static async Task<MobileNetv2Predictor> Create (
+            MLEdgeModel.Configuration configuration = null,
+            string accessKey = null
+        ) {
+            var model = await MLEdgeModel.Create("@natsuite/mobilenet-v2", configuration, accessKey);
+            var predictor = new MobileNetv2Predictor(model);
+            return predictor;
         }
         #endregion
 
@@ -60,7 +88,7 @@ namespace NatML.Vision {
         #region --Operations--
         private readonly MLEdgeModel model;
 
-        void IDisposable.Dispose () { } // Nop
+        private MobileNetv2Predictor (MLEdgeModel model) => this.model = model as MLEdgeModel;
         #endregion
     }
 }