+++ Packages/ai.natml.vision.mobilenet-v2/Changelog.md
@@ -1,3 +1,8 @@
+## 1.0.2
++ Added `MobileNetv2Predictor.Create` static method for creating the predictor.
++ Removed public `MobileNetv2Predictor` constructor.
++ Upgraded to NatML 1.1.3.
+
 ## 1.0.1
 + Upgraded to NatML 1.1.